Update sqlcipher dep to point at upstream.
Signal is currently using a forked version of node-sqlcipher. I'm not
entirely sure as to all the motivations for forking, but it seems that
one reason may have been because upstream didn't support FTS5. This
reason is no longer relevant because of this PR [^1], which adds FTS5
support upstream.

Another reason for the forked version of node-sqlcipher may be to bundle
libcrypto. This is causing problems with many Linux users, especially
those who are not using Ubuntu [^2][^3]. It seems to me that statically
linking is not generally considered to be a best practice on Linux. The
upstream project's README [^4] seems to have a sensible policy: bundling
OpenSSL on Windows and OS X platforms and dynamically linking to the
system's library on Linux platforms.

This PR moves away from the forked node-sqlcipher and returns back to
upstream. This seems to fix OpenSSL problems on non-Ubuntu Linux
distributions.

[^1]: https://github.com/journeyapps/node-sqlcipher/pull/17
[^2]: https://github.com/signalapp/Signal-Desktop/issues/2634
[^3]: https://github.com/signalapp/Signal-Desktop/issues/2662
[^4]: https://github.com/journeyapps/node-sqlcipher#openssl